# cli/openapi/agent-openapi.json
{
  "components": {
    "schemas": {
...
      "ContentDetailResponse": {
        "description": "Detailed response for a single content item.",
        "example": {
          "body_available": true,
          "body_format": "text",
          "body_kind": "article",
          "bullet_points": [
            {
              "category": "key_finding",
              "text": "AI is transforming industries"
            }
          ],
          "can_subscribe": false,
          "content_type": "article",
          "created_at": "2025-06-19T10:30:00Z",
          "display_title": "Understanding AI in 2025",
          "full_markdown": "# Understanding AI in 2025\n\nFull article content...",
          "id": 123,
          "image_url": "/static/images/content/123.png",
          "is_read": false,
          "metadata": {
            "author": "Jane Doe",
            "content_type": "html",
            "publication_date": "2025-06-19T00:00:00Z",
            "source": "Tech Blog",
            "summary": {
              "bullet_points": [
                {
                  "category": "key_finding",
                  "text": "AI is transforming industries"
                }
              ],
              "classification": "to_read",
              "overview": "This article explores the latest developments...",
              "quotes": [
                {
                  "context": "Jane Doe",
                  "text": "The future is now"
                }
              ],
              "summarization_date": "2025-06-19T10:35:00Z",
              "title": "Understanding AI in 2025",
              "topics": [
                "AI",
                "Technology",
                "Future"
              ]
            },
            "word_count": 1500
          },
          "processed_at": "2025-06-19T10:35:00Z",
          "publication_date": "2025-06-18T12:00:00Z",
          "quotes": [
            {
              "context": "Jane Doe",
              "text": "The future is now"
            }
          ],
          "retry_count": 0,
          "short_summary": "This article explores the latest developments...",
          "source": "Tech Blog",
          "status": "completed",
          "structured_summary": {
            "bullet_points": [
              {
                "category": "key_finding",
                "text": "AI is transforming industries"
              }
            ],
            "classification": "to_read",
            "overview": "This article explores the latest developments...",
            "quotes": [
              {
                "context": "Jane Doe",
                "text": "The future is now"
              }
            ],
            "summarization_date": "2025-06-19T10:35:00Z",
            "title": "Understanding AI in 2025",
            "topics": [
              "AI",
              "Technology",
              "Future"
            ]
          },
          "summary": "This article explores the latest developments...",
          "thumbnail_url": "/static/images/thumbnails/123.png",
          "title": "Understanding AI in 2025",
          "topics": [
            "AI",
            "Technology",
            "Future"
          ],
          "updated_at": "2025-06-19T10:35:00Z",
          "url": "https://example.com/article"
        },
        "properties": {
          "body_available": {
            "default": false,
            "description": "Whether canonical body text is available",
            "title": "Body Available",
            "type": "boolean"
          },
          "body_format": {
            "description": "Resolved body format (text or markdown)",
            "nullable": true,
            "title": "Body Format",
            "type": "string"
          },
          "body_kind": {
            "description": "Resolved body kind (article, transcript, or source)",
            "nullable": true,
            "title": "Body Kind",
            "type": "string"
          },
          "bullet_points": {
            "description": "Bullet points from structured summary",
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Bullet Points",
            "type": "array"
          },
          "can_subscribe": {
            "default": false,
            "description": "Whether the current user can subscribe to the detected feed",
            "title": "Can Subscribe",
            "type": "boolean"
          },
          "checked_out_at": {
            "description": "ISO timestamp when content was checked out",
            "nullable": true,
            "title": "Checked Out At",
            "type": "string"
          },
          "checked_out_by": {
            "description": "Worker ID that checked out this content",
            "nullable": true,
            "title": "Checked Out By",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "Type of content (article/podcast/news)"
          },
          "created_at": {
            "description": "ISO timestamp when content was created",
            "title": "Created At",
            "type": "string"
          },
          "detected_feed": {
            "$ref": "#/components/schemas/DetectedFeed",
            "description": "Detected RSS/Atom feed for this content",
            "nullable": true
          },
          "discussion_url": {
            "description": "Discussion URL (tweet, HN thread, etc.) when available",
            "nullable": true,
            "title": "Discussion Url",
            "type": "string"
          },
          "display_title": {
            "description": "Display title (prefers summary title over content title)",
            "title": "Display Title",
            "type": "string"
          },
          "error_message": {
            "description": "Error message if processing failed",
            "nullable": true,
            "title": "Error Message",
            "type": "string"
          },
          "full_markdown": {
            "description": "Full article content formatted as markdown",
            "nullable": true,
            "title": "Full Markdown",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier",
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "description": "URL of full-size AI-generated image for this content",
            "nullable": true,
            "title": "Image Url",
            "type": "string"
          },
          "is_read": {
            "default": false,
            "description": "Whether the content has been marked as read",
            "title": "Is Read",
            "type": "boolean"
          },
          "is_saved_to_knowledge": {
            "default": false,
            "description": "Whether the content has been saved to the user's knowledge library",
            "title": "Is Saved To Knowledge",
            "type": "boolean"
          },
          "metadata": {
            "additionalProperties": true,
            "description": "Content-specific metadata",
            "title": "Metadata",
            "type": "object"
          },
          "news_article_url": {
            "description": "Canonical article link for news content",
            "nullable": true,
            "title": "News Article Url",
            "type": "string"
          },
          "news_discussion_url": {
            "description": "Aggregator discussion URL (HN thread, tweet, etc.)",
            "nullable": true,
            "title": "News Discussion Url",
            "type": "string"
          },
          "news_key_points": {
            "description": "Key points provided for news items",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "title": "News Key Points",
            "type": "array"
          },
          "news_summary": {
            "description": "Short overview synthesized for news items",
            "nullable": true,
            "title": "News Summary",
            "type": "string"
          },
          "processed_at": {
            "description": "ISO timestamp when content was processed",
            "nullable": true,
            "title": "Processed At",
            "type": "string"
          },
          "publication_date": {
            "description": "ISO timestamp of when content was published",
            "nullable": true,
            "title": "Publication Date",
            "type": "string"
          },
          "quotes": {
            "description": "Quotes from structured summary",
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Quotes",
            "type": "array"
          },
          "retry_count": {
            "description": "Number of retry attempts",
            "title": "Retry Count",
            "type": "integer"
          },
          "short_summary": {
            "description": "Short version of summary for list view",
            "nullable": true,
            "title": "Short Summary",
            "type": "string"
          },
          "source": {
            "description": "Content source",
            "nullable": true,
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "description": "Original scraped/submitted URL",
            "nullable": true,
            "title": "Source Url",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus",
            "description": "Processing status"
          },
          "structured_summary": {
            "additionalProperties": true,
            "description": "Structured summary with bullet points and quotes",
            "nullable": true,
            "title": "Structured Summary",
            "type": "object"
          },
          "summary": {
            "description": "Summary text",
            "nullable": true,
            "title": "Summary",
            "type": "string"
          },
          "summary_kind": {
            "$ref": "#/components/schemas/SummaryKind",
            "description": "Summary kind discriminator (e.g., long_interleaved)",
            "nullable": true
          },
          "summary_version": {
            "$ref": "#/components/schemas/SummaryVersion",
            "description": "Summary schema version for the current summary kind",
            "nullable": true
          },
          "thumbnail_url": {
            "description": "URL of 200px thumbnail image for fast loading",
            "nullable": true,
            "title": "Thumbnail Url",
            "type": "string"
          },
          "title": {
            "description": "Content title",
            "nullable": true,
            "title": "Title",
            "type": "string"
          },
          "topics": {
            "description": "Topics from structured summary",
            "items": {
              "type": "string"
            },
            "title": "Topics",
            "type": "array"
          },
          "updated_at": {
            "description": "ISO timestamp of last update",
            "nullable": true,
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "description": "Canonical URL of the content",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content_type",
          "url",
          "display_title",
          "status",
          "retry_count",
          "metadata",
          "created_at",
          "bullet_points",
          "quotes",
          "topics"
        ],
        "title": "ContentDetailResponse",
        "type": "object"
      },
...
      "ContentListResponse": {
        "description": "Response for content list endpoint.",
        "example": {
          "available_dates": [
            "2025-06-19",
            "2025-06-18"
          ],
          "content_types": [
            "article",
            "podcast",
            "news"
          ],
          "contents": [
            {
              "classification": "to_read",
              "content_type": "article",
              "created_at": "2025-06-19T10:30:00Z",
              "id": 123,
              "platform": "substack",
              "processed_at": "2025-06-19T10:35:00Z",
              "short_summary": "This article explores...",
              "source": "Tech Blog",
              "status": "completed",
              "title": "Understanding AI in 2025",
              "url": "https://example.com/article"
            }
          ],
          "meta": {
            "has_more": true,
            "next_cursor": "eyJsYXN0X2lkIjoxMjN9",
            "page_size": 25,
            "total": 1
          }
        },
        "properties": {
          "available_dates": {
            "description": "List of available dates (YYYY-MM-DD)",
            "items": {
              "type": "string"
            },
            "title": "Available Dates",
            "type": "array"
          },
          "content_types": {
            "description": "Available content types for filtering",
            "items": {
              "$ref": "#/components/schemas/ContentType"
            },
            "title": "Content Types",
            "type": "array"
          },
          "contents": {
            "description": "List of content items",
            "items": {
              "$ref": "#/components/schemas/ContentSummaryResponse"
            },
            "title": "Contents",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetadata",
            "description": "Pagination metadata for the response"
          }
        },
        "required": [
          "contents",
          "available_dates",
          "content_types",
          "meta"
        ],
        "title": "ContentListResponse",
        "type": "object"
      },
...
      "ContentType": {
        "description": "Supported content types in storage and API contracts.",
        "enum": [
          "article",
          "podcast",
          "news",
          "insight_report",
          "unknown"
        ],
        "title": "ContentType",
        "type": "string"
      },
...
      "DetectedFeed": {
        "description": "Detected RSS/Atom feed from content page.",
        "example": {
          "format": "rss",
          "title": "Example Newsletter",
          "type": "substack",
          "url": "https://example.substack.com/feed"
        },
        "properties": {
          "format": {
            "default": "rss",
            "description": "Feed format: rss or atom",
            "title": "Format",
            "type": "string"
          },
          "title": {
            "description": "Feed title from link tag",
            "nullable": true,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Feed type: substack, podcast_rss, or atom",
            "title": "Type",
            "type": "string"
          },
          "url": {
            "description": "Feed URL",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "type"
        ],
        "title": "DetectedFeed",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginationMetadata": {
        "description": "Pagination metadata for responses.",
        "properties": {
          "has_more": {
            "default": false,
            "description": "Whether more results are available",
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "description": "Opaque cursor token for next page (null if no more results)",
            "nullable": true,
            "title": "Next Cursor",
            "type": "string"
          },
          "page_size": {
            "description": "Number of items in current response",
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total number of items (expensive to compute, may be omitted)",
            "nullable": true,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page_size"
        ],
        "title": "PaginationMetadata",
        "type": "object"
      },
...
      "SubmissionStatusListResponse": {
        "description": "Response for user submission status list.",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMetadata",
            "description": "Pagination metadata for the response"
          },
          "submissions": {
            "description": "List of user-submitted items still processing or failed",
            "items": {
              "$ref": "#/components/schemas/SubmissionStatusResponse"
            },
            "title": "Submissions",
            "type": "array"
          }
        },
        "required": [
          "submissions",
          "meta"
        ],
        "title": "SubmissionStatusListResponse",
        "type": "object"
      },
      "SubmissionStatusResponse": {
        "description": "Status information for a user-submitted content item.",
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "Type of content (article/podcast/news/unknown)"
          },
          "created_at": {
            "description": "ISO timestamp when content was created",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "description": "Failure reason when status=failed/skipped",
            "nullable": true,
            "title": "Error Message",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier",
            "title": "Id",
            "type": "integer"
          },
          "is_self_submission": {
            "default": true,
            "description": "Whether this content was submitted by the current user",
            "title": "Is Self Submission",
            "type": "boolean"
          },
          "processed_at": {
            "description": "ISO timestamp when content was processed",
            "nullable": true,
            "title": "Processed At",
            "type": "string"
          },
          "source_url": {
            "description": "Original submitted URL",
            "nullable": true,
            "title": "Source Url",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus",
            "description": "Processing status"
          },
          "submitted_via": {
            "description": "Submission channel (share_sheet, etc.)",
            "nullable": true,
            "title": "Submitted Via",
            "type": "string"
          },
          "title": {
            "description": "Content title (if detected)",
            "nullable": true,
            "title": "Title",
            "type": "string"
          },
          "url": {
            "description": "Canonical URL of the content",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content_type",
          "url",
          "status",
          "created_at"
        ],
        "title": "SubmissionStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Filtered machine-oriented API contract for the standalone newsly-agent CLI.",
    "title": "Newsly Agent CLI API",
    "version": "2.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/content/": {
      "get": {
        "description": "Retrieve a list of content items with optional filtering by content type and date. Supports multiple content types via repeated query parameters (e.g., ?content_type=article&content_type=podcast). Supports cursor-based pagination for efficient loading.",
        "operationId": "listContents",
        "parameters": [
          {
            "description": "Filter by content type(s). Can be specified multiple times for multiple types (article/podcast/news)",
            "in": "query",
            "name": "content_type",
            "required": false,
            "schema": {
              "description": "Filter by content type(s). Can be specified multiple times for multiple types (article/podcast/news)",
              "items": {
                "type": "string"
              },
              "nullable": true,
              "title": "Content Type",
              "type": "array"
            }
          },
          {
            "description": "Filter by date (YYYY-MM-DD format)",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "description": "Filter by date (YYYY-MM-DD format)",
              "nullable": true,
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "description": "Filter by read status (all/read/unread)",
            "in": "query",
            "name": "read_filter",
            "required": false,
            "schema": {
              "default": "all",
              "description": "Filter by read status (all/read/unread)",
              "pattern": "^(all|read|unread)$",
              "title": "Read Filter",
              "type": "string"
            }
          },
          {
            "description": "Pagination cursor for next page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Pagination cursor for next page",
              "nullable": true,
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "description": "Number of items per page (max 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Number of items per page (max 100)",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Whether to include the available_dates metadata used by filterable list UIs. Disable for feed surfaces that do not show date filters.",
            "in": "query",
            "name": "include_available_dates",
            "required": false,
            "schema": {
              "default": true,
              "description": "Whether to include the available_dates metadata used by filterable list UIs. Disable for feed surfaces that do not show date filters.",
              "title": "Include Available Dates",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List content items",
        "tags": [
          "content"
        ]
      }
    },
    "/api/content/submissions/list": {
      "get": {
        "description": "Returns self-submitted content items that are not yet completed, including processing, failed, and skipped statuses.",
        "operationId": "listContentSubmissionStatuses",
        "parameters": [
          {
            "description": "Pagination cursor for next page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Pagination cursor for next page",
              "nullable": true,
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "description": "Number of items per page (max 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Number of items per page (max 100)",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionStatusListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List user-submitted content still processing or failed",
        "tags": [
          "content"
        ]
      }
    },
...
    "/api/content/{content_id}": {
      "get": {
        "description": "Retrieve detailed information about a specific content item.",
        "operationId": "getContentDetail",
        "parameters": [
          {
            "description": "Content ID",
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "description": "Content ID",
              "exclusiveMinimum": true,
              "minimum": 0,
              "title": "Content Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Content not found"
                }
              }
            },
            "description": "Content not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get content details",
        "tags": [
          "content"
        ]
      }
    },
...
  },
...
}